fix(connectors): respect the function deny-list in Postgres/ClickHouse/ODBC federation
Generalize the Snowflake deny-list federation wrapper and apply it to every
SqlTable-backed federation connector, so Spice-only UDFs (json_get_str, the
embedding/distance UDFs, etc.) are evaluated locally instead of being unparsed
into the SQL sent to the remote engine (which rejects them). Same bug class as
#10703, which was previously only fixed for Snowflake/MySQL/DuckDB.

Root cause: upstream `impl SQLExecutor for SqlTable<T,P>` inherits the default
`can_execute_plan` (always federate) and ignores `function_support`. MySQLTable/
SQLiteTable/DuckDBTable override it; SqlTable-backed connectors (Postgres,
ClickHouse, ODBC, Snowflake) do not.

- data_components: lift `DenyFunctionsSqlExecutor` + `create_spice_federated_table_provider`
  out of the Snowflake module into a shared `data_components::federation` module
  (Snowflake re-exports it; behaviour unchanged). Gated behind a new `federation`
  feature that pulls `datafusion-table-providers/federation` (the upstream feature
  that provides `impl SQLExecutor for SqlTable`).
- ClickHouse / ODBC: `ClickhouseTableFactory` and `ODBCTableFactory` now carry an
  optional `FunctionSupport` (`with_function_support`) and build their provider via
  the shared wrapper instead of upstream's `create_federated_table_provider`.
- Postgres: `connector-postgres` builds its read provider via a local
  `federated_postgres_table_provider` that mirrors the upstream factory internals
  but routes pushdown through the shared wrapper.
- All three connectors install `deny_spice_specific_functions()`.

Connectors that don't use SQL federation (MSSQL, Oracle, Spark, Databricks,
ScyllaDB, MongoDB) push no function SQL down, so there's nothing to deny there.
FlightSQL/Dremio (custom Flight federation) are handled separately.

Tests: the shared `federation` module carries the wrapper unit tests (moved from
Snowflake): can_execute_plan blocks/allows by deny-list, and the adaptor wires the
deny-list executor. Live end-to-end federation for these connectors is covered by
the existing CI integration suites (they require running Postgres/ClickHouse/ODBC
servers).

Refs #10703
