+++ crates/data-connectors/connector-odbc/src/lib.rs
@@ -29,6 +29,7 @@ use runtime::dataconnector::{
     ConnectorComponent, ConnectorParams, DataConnector, DataConnectorFactory, DataConnectorResult,
     NewDataConnectorResult,
 };
+use runtime::datafusion::udf::deny_spice_specific_functions;
 use runtime::parameters::{ParameterSpec, Parameters};
 use snafu::prelude::*;
 use std::any::Any;
@@ -267,7 +268,11 @@ impl DataConnectorFactory for ODBCFactory {
                     .context(UnableToCreateODBCConnectionPoolSnafu)?,
             );
 
-            let odbc_factory = ODBCTableFactory::new(pool, dialect);
+            // Install the Spice function deny-list so Spice-only UDFs
+            // (json_get_str, etc.) are evaluated locally instead of pushed into
+            // the SQL sent through ODBC, which would reject them (#10703).
+            let odbc_factory = ODBCTableFactory::new(pool, dialect)
+                .with_function_support(deny_spice_specific_functions().as_ref().clone());
 
             Ok(Arc::new(ODBC { odbc_factory }) as Arc<dyn DataConnector>)
         })
