+++ crates/data_components/Cargo.toml
@@ -128,7 +128,11 @@ rdkafka = { workspace = true, features = ["cmake-build"], optional = true }
 [features]
 bench = [] # Feature for benchmarking that exposes internal functions
 adbc = ["datafusion-table-providers/adbc-federation"]
-clickhouse = ["dep:clickhouse-rs", "datafusion-table-providers/federation"]
+# Shared Spice federation deny-list wrapper (crate::federation). Requires the
+# upstream `impl SQLExecutor for SqlTable`, which is gated behind
+# `datafusion-table-providers/federation`.
+federation = ["datafusion-table-providers/federation"]
+clickhouse = ["dep:clickhouse-rs", "federation"]
 cosmosdb = ["dep:azure_core", "dep:azure_data_cosmos"]
 databricks = [
     "delta_lake",
@@ -162,13 +166,14 @@ mysql = [
     "datafusion-table-providers/mysql",
     "datafusion-table-providers/mysql-federation",
 ]
-odbc = ["datafusion-table-providers/federation", "db_connection_pool/odbc"]
+odbc = ["federation", "db_connection_pool/odbc"]
 oracle = ["dep:oracle", "dep:bb8", "dep:bb8-oracle", "dep:bigdecimal"]
 postgres = [
     "dep:tokio-postgres",
     "dep:pgwire-replication",
     "dep:native-tls",
     "dep:postgres-native-tls",
+    "federation",
     "datafusion-table-providers/postgres",
     "datafusion-table-providers/postgres-federation",
 ]
@@ -178,7 +183,7 @@ sharepoint = ["dep:graph-rs-sdk"]
 # Enables graph-rs-sdk's `test-util` feature so `GraphClient::use_test_endpoint`
 # accepts non-Microsoft hosts (e.g. localhost mocks). Intended for tests only.
 sharepoint-mock-host = ["sharepoint", "graph-rs-sdk/test-util"]
-snowflake = ["dep:snowflake-api", "datafusion-table-providers/federation", "db_connection_pool/snowflake"]
+snowflake = ["dep:snowflake-api", "federation", "db_connection_pool/snowflake"]
 spark_connect = ["dep:spark-connect-rs"]
 sqlite = [
     "dep:rusqlite",
