+++ crates/data_components/src/federation.rs
@@ -1,0 +1,306 @@
+/*
+Copyright 2024-2025 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+//! Spice-owned `SQLExecutor` wrapper that applies a function deny-list to
+//! federation pushdown decisions for any [`SqlTable`]-backed connector.
+//!
+//! Upstream `impl SQLExecutor for SqlTable<T, P>` inherits the trait default for
+//! `can_execute_plan`, which always returns `true`. As a result, any plan —
+//! including ones with Spice-only UDFs like `json_get_str` — is pushed to the
+//! remote engine, where it fails. This wrapper consults a `FunctionSupport`
+//! deny-list and refuses federation when an unsupported function appears, so
+//! `DataFusion` evaluates the affected expression locally instead.
+//!
+//! Connectors backed by a `SqlTable` (Snowflake, `ClickHouse`, `ODBC`,
+//! `PostgreSQL`, …) build their federated provider with
+//! [`create_spice_federated_table_provider`] instead of upstream's
+//! `SqlTable::create_federated_table_provider`. (`MySQL`, `SQLite`, and `DuckDB`
+//! don't need this wrapper — their upstream table types already consult
+//! `function_support` in `can_execute_plan`.)
+
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use datafusion::{
+    arrow::datatypes::SchemaRef,
+    error::Result as DataFusionResult,
+    logical_expr::LogicalPlan,
+    physical_plan::{PhysicalExpr, SendableRecordBatchStream},
+    sql::{TableReference, unparser::dialect::Dialect},
+};
+use datafusion_federation::{
+    FederatedTableProviderAdaptor, FederatedTableSource,
+    sql::{RemoteTableRef, SQLExecutor, SQLFederationProvider, SQLTableSource},
+};
+use datafusion_table_providers::sql::sql_provider_datafusion::SqlTable;
+use datafusion_table_providers::util::supported_functions::{
+    FunctionSupport, contains_unsupported_functions,
+};
+
+/// `SQLExecutor` that delegates to a wrapped [`SqlTable`] but overrides
+/// [`SQLExecutor::can_execute_plan`] to consult a function deny-list.
+pub struct DenyFunctionsSqlExecutor<T: 'static, P: 'static> {
+    inner: Arc<SqlTable<T, P>>,
+    function_support: Option<FunctionSupport>,
+}
+
+impl<T: 'static, P: 'static> DenyFunctionsSqlExecutor<T, P> {
+    #[must_use]
+    pub fn new(inner: Arc<SqlTable<T, P>>, function_support: Option<FunctionSupport>) -> Self {
+        Self {
+            inner,
+            function_support,
+        }
+    }
+}
+
+impl<T, P> std::fmt::Debug for DenyFunctionsSqlExecutor<T, P> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("DenyFunctionsSqlExecutor")
+            .field("inner", &self.inner)
+            .field("function_support", &self.function_support)
+            .finish()
+    }
+}
+
+#[async_trait]
+impl<T: 'static, P: 'static> SQLExecutor for DenyFunctionsSqlExecutor<T, P> {
+    fn name(&self) -> &str {
+        SQLExecutor::name(self.inner.as_ref())
+    }
+
+    fn compute_context(&self) -> Option<String> {
+        SQLExecutor::compute_context(self.inner.as_ref())
+    }
+
+    fn dialect(&self) -> Arc<dyn Dialect> {
+        SQLExecutor::dialect(self.inner.as_ref())
+    }
+
+    fn can_execute_plan(&self, plan: &LogicalPlan) -> bool {
+        // Default to allowing federation when no deny-list is configured. When
+        // one is provided, reject any plan that touches a denied function so
+        // DataFusion evaluates it locally instead of pushing unparseable SQL to
+        // the remote engine.
+        self.function_support.as_ref().is_none_or(|func_supp| {
+            !contains_unsupported_functions(plan, func_supp).unwrap_or(false)
+        })
+    }
+
+    fn execute(
+        &self,
+        query: &str,
+        schema: SchemaRef,
+        filters: &[Arc<dyn PhysicalExpr>],
+    ) -> DataFusionResult<SendableRecordBatchStream> {
+        SQLExecutor::execute(self.inner.as_ref(), query, schema, filters)
+    }
+
+    async fn table_names(&self) -> DataFusionResult<Vec<String>> {
+        SQLExecutor::table_names(self.inner.as_ref()).await
+    }
+
+    async fn get_table_schema(&self, table_name: &str) -> DataFusionResult<SchemaRef> {
+        SQLExecutor::get_table_schema(self.inner.as_ref(), table_name).await
+    }
+}
+
+/// Build a [`FederatedTableProviderAdaptor`] for a [`SqlTable`] that routes
+/// plan-pushdown decisions through [`DenyFunctionsSqlExecutor`].
+///
+/// This mirrors [`SqlTable::create_federated_table_provider`] but installs the
+/// wrapper executor so that any plan referencing a function on the supplied
+/// `function_support` deny-list (e.g. `json_get_str`) falls back to local
+/// `DataFusion` evaluation instead of being unparsed into the remote engine's
+/// SQL.
+#[must_use]
+pub fn create_spice_federated_table_provider<T: 'static, P: 'static>(
+    table: Arc<SqlTable<T, P>>,
+    schema: SchemaRef,
+    table_reference: TableReference,
+    function_support: Option<FunctionSupport>,
+) -> FederatedTableProviderAdaptor {
+    let executor: Arc<dyn SQLExecutor> = Arc::new(DenyFunctionsSqlExecutor::new(
+        Arc::clone(&table),
+        function_support,
+    ));
+    let fed_provider = Arc::new(SQLFederationProvider::new(executor));
+    let table_source: Arc<dyn FederatedTableSource> = Arc::new(SQLTableSource::new_with_schema(
+        fed_provider,
+        RemoteTableRef::from(table_reference),
+        schema,
+    ));
+    FederatedTableProviderAdaptor::new_with_provider(table_source, table)
+}
+
+#[cfg(test)]
+mod tests {
+    use std::any::Any;
+    use std::error::Error;
+
+    use async_trait::async_trait;
+    use datafusion::arrow::datatypes::{DataType, Field, Schema};
+    use datafusion::datasource::TableProvider;
+    use datafusion::logical_expr::{
+        ColumnarValue, Expr, LogicalPlan, LogicalPlanBuilder, ScalarUDF, TableSource, Volatility,
+        builder::LogicalTableSource, create_udf, expr::ScalarFunction,
+    };
+    use datafusion::prelude::col;
+    use datafusion_federation::sql::SQLExecutor;
+    use datafusion_table_providers::sql::db_connection_pool::{
+        DbConnectionPool, JoinPushDown, dbconnection::DbConnection,
+    };
+    use datafusion_table_providers::util::supported_functions::{
+        FunctionRestriction, FunctionSupport,
+    };
+
+    use super::*;
+
+    struct MockConn {}
+
+    impl DbConnection<(), &'static dyn ToString> for MockConn {
+        fn as_any(&self) -> &dyn Any {
+            self
+        }
+
+        fn as_any_mut(&mut self) -> &mut dyn Any {
+            self
+        }
+    }
+
+    struct MockPool {}
+
+    #[async_trait]
+    impl DbConnectionPool<(), &'static dyn ToString> for MockPool {
+        async fn connect(
+            &self,
+        ) -> Result<Box<dyn DbConnection<(), &'static dyn ToString>>, Box<dyn Error + Send + Sync>>
+        {
+            Ok(Box::new(MockConn {}))
+        }
+
+        fn join_push_down(&self) -> JoinPushDown {
+            JoinPushDown::Disallow
+        }
+    }
+
+    fn test_sql_table() -> Arc<SqlTable<(), &'static dyn ToString>> {
+        let pool: Arc<dyn DbConnectionPool<(), &'static dyn ToString> + Send + Sync> =
+            Arc::new(MockPool {});
+        let schema = Arc::new(Schema::new(vec![
+            Field::new("id", DataType::Int32, false),
+            Field::new("val", DataType::Utf8, true),
+        ]));
+        Arc::new(SqlTable::new_with_schema(
+            "test",
+            &pool,
+            schema,
+            TableReference::bare("t"),
+            None,
+        ))
+    }
+
+    fn stub_udf(name: &str) -> Arc<ScalarUDF> {
+        Arc::new(create_udf(
+            name,
+            vec![DataType::Utf8],
+            DataType::Utf8,
+            Volatility::Immutable,
+            Arc::new(|args: &[ColumnarValue]| Ok(args[0].clone())),
+        ))
+    }
+
+    fn scan_with_projection(udf_name: &str) -> LogicalPlan {
+        let schema = Arc::new(Schema::new(vec![
+            Field::new("id", DataType::Int32, false),
+            Field::new("val", DataType::Utf8, true),
+        ]));
+        let source = Arc::new(LogicalTableSource::new(schema)) as Arc<dyn TableSource>;
+        LogicalPlanBuilder::scan("t", source, None)
+            .expect("scan")
+            .project(vec![Expr::ScalarFunction(ScalarFunction::new_udf(
+                stub_udf(udf_name),
+                vec![col("val")],
+            ))])
+            .expect("project")
+            .build()
+            .expect("build")
+    }
+
+    fn deny_support(names: &[&str]) -> FunctionSupport {
+        FunctionSupport::new(
+            Some(FunctionRestriction::Deny(
+                names.iter().map(|s| (*s).to_string()).collect(),
+            )),
+            None,
+            None,
+        )
+    }
+
+    #[test]
+    fn can_execute_plan_allows_when_no_deny_list() {
+        let executor = DenyFunctionsSqlExecutor::new(test_sql_table(), None);
+        let plan = scan_with_projection("json_get_str");
+        assert!(
+            executor.can_execute_plan(&plan),
+            "no deny-list should permit federation of any plan"
+        );
+    }
+
+    #[test]
+    fn can_execute_plan_blocks_denied_function() {
+        let executor =
+            DenyFunctionsSqlExecutor::new(test_sql_table(), Some(deny_support(&["json_get_str"])));
+        let plan = scan_with_projection("json_get_str");
+        assert!(
+            !executor.can_execute_plan(&plan),
+            "deny-listed function in projection must block federation"
+        );
+    }
+
+    #[test]
+    fn can_execute_plan_allows_non_denied_function() {
+        let executor =
+            DenyFunctionsSqlExecutor::new(test_sql_table(), Some(deny_support(&["json_get_str"])));
+        let plan = scan_with_projection("upper");
+        assert!(
+            executor.can_execute_plan(&plan),
+            "function not on deny-list should still federate"
+        );
+    }
+
+    #[test]
+    fn create_spice_federated_table_provider_wires_deny_list() {
+        // Regression test for #10703: building the federated adaptor must
+        // route can_execute_plan decisions through the deny-list wrapper, not
+        // upstream's default `true`-everywhere impl.
+        let table = test_sql_table();
+        let schema = table.schema();
+        let adaptor = create_spice_federated_table_provider(
+            Arc::clone(&table),
+            Arc::clone(&schema),
+            TableReference::bare("t"),
+            Some(deny_support(&["json_get_str"])),
+        );
+
+        // The adaptor's TableProvider fallback must be the underlying SqlTable
+        // so non-federation contexts still scan correctly.
+        assert!(adaptor.table_provider.is_some(), "fallback provider wired");
+
+        // Federation source schema must match the SqlTable schema verbatim.
+        assert_eq!(adaptor.source.schema().as_ref(), schema.as_ref());
+    }
+}
